fix: register club BasicInformFrame select component refactoring
